timetable/manager: Add TimetableManager cls
The MRTAFactory creates one instance of the TimetableManager and passes it to the components (auctioneer, dispatcher) that need it
The MRTAFactory receives the fleet (list of robot_ids). The TimetableManager uses this information to initialize the timetables of all robots in the fleet
The ccu has a TimetableManager that includes timetables for all robots in the fleet.
The robot has a TimetableManager that includes only the timetable of the robot (no information of other robots is included)