Fix alternative timeslot computation. Accept suggested timeslot
Before, the program crashed before computing alternative timeslots. Now, the task is added with loose temporal constraints in each possible insertion point.
Times are relative to the zero timepoint and converted back to global times.
Before, the alternative timeslot was added to a dict 'waiting for user confirmation' and no other action was taken. Now, the suggestion is always accepted
(Provisional fix until the feature for asking for user confirmation is added).